Separate out db.go
